Move paths file to test package
