+++ multisensor_pipeline/tests/paths.py
