+++ multisensor_pipeline/tests/test_webcam.py
@@ -6,7 +6,7 @@ from time import sleep
 
 import av
 import pytest
-from multisensor_pipeline.modules.paths import DATA_PATH
+from multisensor_pipeline.tests.paths import DATA_PATH
 
 from multisensor_pipeline.modules import QueueSink
 from multisensor_pipeline.modules.video.webcam import WebCamSource
